Test `formatDate` function

+++ src/utils/misc.js
@@ -1,6 +1,6 @@
 const formatDate = date =>
-  new Intl.DateTimeFormat('en-US', {month: 'short', year: '2-digit'}).format(
+  new Intl.DateTimeFormat('en-US', { month: 'long', year: '2-digit' }).format(
     date,
   )
 
-export {formatDate}
+export { formatDate }
